refactor(sandbox-fc): extract state and snapshot restore helpers
